fix: set configPaths options to empty object
The configPaths option can be empty by default and use the default
attributes for tsconfig.json or jsconfig.json. Thus the options
object should be allowed to be undefined so that we don't overload
the config with an empty object if we don't need to define another
path for ts/jsonconfig.